Move shufflings code together into seed util
